+++ log.go
@@ -69,11 +69,11 @@ func (g *game) PrintEntry(e logEntry) {
 }
 
 func (g *game) StoryPrint(s string) {
-	g.Stats.Story = append(g.Stats.Story, fmt.Sprintf("Depth %2d|Turn %7.1f| %s", g.Depth, float64(g.Turn)/10, s))
+	g.Stats.Story = append(g.Stats.Story, fmt.Sprintf("Depth %2d|Turn %5d| %s", g.Depth, g.Turn/10, s))
 }
 
 func (g *game) StoryPrintf(format string, a ...interface{}) {
-	g.Stats.Story = append(g.Stats.Story, fmt.Sprintf("Depth %2d|Turn %7.1f| %s", g.Depth, float64(g.Turn)/10, fmt.Sprintf(format, a...)))
+	g.Stats.Story = append(g.Stats.Story, fmt.Sprintf("Depth %2d|Turn %5d| %s", g.Depth, g.Turn/10, fmt.Sprintf(format, a...)))
 }
 
 func (g *game) CrackSound() (text string) {
